requirements.txt
astroid==2.15.5
autopep8==2.0.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
docstring-to-markdown==0.12
flake8==6.0.0
fonttools==4.39.4
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.12.0
jedi==0.18.2
kiwisolver==1.4.4
langdetect==1.0.9
lazy-object-proxy==1.9.0
libretranslatepy==2.1.1
lxml==4.9.2
matplotlib==3.7.1
mccabe==0.7.0
msgpack==1.0.5
mypy==1.3.0
mypy-extensions==1.0.0
neovim==0.3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
pylint==2.17.4
pyls-isort==0.2.2
pylsp-mypy==0.6.7
pynvim==0.4.3
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.3
pytoolconfig==1.2.5
requests==2.31.0
rfc3986==1.5.0
rope==1.8.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
tomli==2.0.1
tomlkit==0.11.8
translate==3.6.1
typing_extensions==4.6.3
ujson==5.7.0
urllib3==2.0.3
whatthepatch==1.0.5
wrapt==1.15.0
yapf==0.33.0
